Force CameraController to unbind all when changing UseCases combination
Force unbind all to recalculate the resolutions for the new UseCases set. This can avoid the issue that when previously a UseCase may select MAXIMUM resolution to disallow additional UseCases being bound.

Relnote: "Fixed a bug in CameraController that caused an IllegalStateException when an initial UseCase selected the maximum resolution, which subsequently prevented other UseCases from being bound."

Bug: 440374234
Test: CameraControllerFragmentTest
Change-Id: Ifb758bef1582b20717a64b67657a9358e5600034